Bug 1668926 - fix unintentional integer overflow in nsMsgDBFolder::HandleAutoCompactEvent. r=benc
CID 1467532 (#1 of 1): Unintentional integer overflow (OVERFLOW_BEFORE_WIDEN)overflow_before_widen: Potentially overflowing expression purgeThreshold * 1024 with type int (32 bits, signed) is evaluated using 32-bit arithmetic, and then used in a context that expects an expression of type int64_t (64 bits, signed).

Differential Revision: https://phabricator.services.mozilla.com/D111616

--HG--
extra : rebase_source : e9e7189ea59b7685a876e4f7222380065f409910

+++ mailnews/base/src/nsMsgDBFolder.cpp
@@ -1697,7 +1697,7 @@ nsresult nsMsgDBFolder::HandleAutoCompactEvent(nsIMsgWindow* aWindow) {
       int32_t purgeThreshold;
       rv = GetPurgeThreshold(&purgeThreshold);
       NS_ENSURE_SUCCESS(rv, rv);
-      if (totalExpungedBytes > (purgeThreshold * 1024)) {
+      if (totalExpungedBytes > ((int64_t)purgeThreshold * 1024)) {
         bool okToCompact = false;
         nsCOMPtr<nsIPrefService> pref =
             do_GetService(NS_PREFSERVICE_CONTRACTID);
